WEB/HW11_new/pyproject.toml
[tool.poetry]
name = "hw11-new"
version = "0.1.0"
description = ""
authors = ["Kagev <kaplitskiygv@gmail.com>"]
readme = "README.md"
packages = [{include = "hw11_new"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
faker = "^19.6.0"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"
pydantic = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
